api/requirements.txt
Django
djangorestframework
Pillow
django-cors-headers
django-filter
pylint
django-import-export
tablib